+++ lib/class/search.class.php
@@ -430,6 +430,21 @@ class Search extends playlist_object
     }
 
     /**
+     * has_image
+     *
+     * Image Exists? (Album, Artist)
+     */
+    private function has_image()
+    {
+        $this->types[] = array(
+            'name' => 'has image',
+            'label' => T_('Local Image'),
+            'type' => 'boolean',
+            'widget' => array('input', 'hidden')
+        );
+    }
+
+    /**
      * image_height
      *
      * Image Height (Album, Artist)
@@ -912,9 +927,10 @@ class Search extends playlist_object
         $this->last_play();
         $this->total_time();
         $this->played_times();
+        $this->has_image();
         $this->image_width();
         $this->image_height();
-    }
+    } // artisttypes
 
     /**
      * albumtypes
@@ -988,10 +1004,10 @@ class Search extends playlist_object
             'type' => 'boolean_numeric',
             'widget' => array('select', $catalogs)
         );
-
+        $this->has_image();
         $this->image_width();
         $this->image_height();
-    }
+    } // albumtypes
 
     /**
      * videotypes
@@ -1666,6 +1682,10 @@ class Search extends playlist_object
                         $join['tag'][$key] = "find_in_set('$input', cast(`realtag_$key`.`name` as char)) $sql_match_operator 0";
                     }
                 break;
+                case 'has image':
+                    $where[]        = ($sql_match_operator == '1') ? "`image`.`object_id` IS NOT NULL" : "`has_image`.`object_id` IS NULL";
+                    $table['has_image'] = "LEFT JOIN (SELECT `object_id` from `image` WHERE `object_type` = 'album') as `has_image` ON `album`.`id` = `has_image`.`object_id`";
+                break;
                 case 'image height':
                     $where[]       = "`image`.`height` $sql_match_operator '$input'";
                     $join['image'] = true;
@@ -1837,6 +1857,10 @@ class Search extends playlist_object
                                 "' ', `artist`.`name`)) $sql_match_operator '$input') AND `user_flag`.`user` = $userid";
                     $where[] .= "`user_flag`.`object_type` = 'artist'";
                 break;
+                case 'has image':
+                    $where[]            = ($sql_match_operator == '1') ? "`has_image`.`object_id` IS NOT NULL" : "`has_image`.`object_id` IS NULL";
+                    $table['has_image'] = "LEFT JOIN (SELECT `object_id` from `image` WHERE `object_type` = 'artist') as `has_image` ON `artist`.`id` = `has_image`.`object_id`";
+                    break;
                 case 'image height':
                     $where[]       = "`image`.`height` $sql_match_operator '$input'";
                     $join['image'] = true;
@@ -1993,9 +2017,6 @@ class Search extends playlist_object
             $raw_input          = $this->_mangle_data($rule[2], $type, $operator);
             $input              = filter_var($raw_input, FILTER_SANITIZE_STRING, FILTER_FLAG_NO_ENCODE_QUOTES);
             $sql_match_operator = $operator['sql'];
-            $tag_join           = '';
-            $album_tag_join     = '';
-            $artist_tag_join    = '';
 
             switch ($rule[0]) {
                 case 'anywhere':
@@ -2299,13 +2320,13 @@ class Search extends playlist_object
                     $where[]       = "`song`.`update_time` $sql_match_operator $input";
                     break;
                 case 'recent_added':
-                    $key = md5($input . $sql_match_operator);
-                    $where[] = "`addition_time_$key`.`id` IS NOT NULL";
+                    $key                       = md5($input . $sql_match_operator);
+                    $where[]                   = "`addition_time_$key`.`id` IS NOT NULL";
                     $table['addition_' . $key] = "LEFT JOIN (SELECT `id` from `song` ORDER BY $sql_match_operator DESC LIMIT $input) as `addition_time_$key` ON `song`.`id` = `addition_time_$key`.`id`";
                     break;
                 case 'recent_updated':
-                    $key = md5($input . $sql_match_operator);
-                    $where[] = "`update_time_$key`.`id` IS NOT NULL";
+                    $key                     = md5($input . $sql_match_operator);
+                    $where[]                 = "`update_time_$key`.`id` IS NOT NULL";
                     $table['update_' . $key] = "LEFT JOIN (SELECT `id` from `song` ORDER BY $sql_match_operator DESC LIMIT $input) as `update_time_$key` ON `song`.`id` = `update_time_$key`.`id`";
                     break;
                 case 'metadata':